# docs/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Tasks.Api",
        "version": "v1"
    },
    "paths": {
...
        "/Sections/{id}": {
            "put": {
                "tags": [
                    "Sections"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "style": "simple",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateSectionRequest"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateSectionRequest"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateSectionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/SectionDto"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SectionDto"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SectionDto"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Sections"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "style": "simple",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "get": {
                "tags": [
                    "Sections"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "style": "simple",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/SectionDto"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SectionDto"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SectionDto"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/Subtasks/{id}": {
            "delete": {
                "tags": [
                    "Subtasks"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "style": "simple",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
        "/Tasks/{id}": {
            "put": {
                "tags": [
                    "Tasks"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "style": "simple",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTaskRequest"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTaskRequest"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTaskRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskDto"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskDto"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskDto"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "Tasks"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "style": "simple",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskDto"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskDto"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskDto"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "CommentDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "text": {
                        "type": "string",
                        "nullable": true
                    },
                    "createdAtUtc": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "createdBy": {
                        "$ref": "#/components/schemas/UserProfileDto"
                    }
                },
                "additionalProperties": false
            },
...
            "SectionDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "projectId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string",
                        "nullable": true
                    },
                    "tasks": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TaskDto"
                        },
                        "nullable": true
                    }
                },
                "additionalProperties": false
            },
            "Subtask": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "title": {
                        "type": "string",
                        "nullable": true
                    },
                    "isComplete": {
                        "type": "boolean"
                    },
                    "createdAtUtc": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "completedAtUtc": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true,
                        "readOnly": true
                    },
                    "parentId": {
                        "type": "string",
                        "format": "uuid"
                    }
                },
                "additionalProperties": false
            },
            "TagDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string",
                        "nullable": true
                    },
                    "color": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "additionalProperties": false
            },
            "TaskDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "title": {
                        "type": "string",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "priority": {
                        "$ref": "#/components/schemas/TaskPriority"
                    },
                    "isComplete": {
                        "type": "boolean"
                    },
                    "createdAtUtc": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "completedAtUtc": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "dueAtUtc": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "projectId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "sectionId": {
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                    },
                    "createdBy": {
                        "$ref": "#/components/schemas/UserProfileDto"
                    },
                    "assignedTo": {
                        "$ref": "#/components/schemas/UserProfileDto"
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TagDto"
                        },
                        "nullable": true
                    },
                    "subtasks": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Subtask"
                        },
                        "nullable": true
                    },
                    "comments": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CommentDto"
                        },
                        "nullable": true
                    }
                },
                "additionalProperties": false
            },
            "TaskPriority": {
                "enum": [
                    0,
                    1,
                    2,
                    3,
                    4
                ],
                "type": "integer",
                "format": "int32"
            },
...
            "UpdateSectionRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "nullable": true
                    },
                    "orderIndex": {
                        "type": "integer",
                        "format": "int32"
                    }
                },
                "additionalProperties": false
            },
...
            "UpdateTaskRequest": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "priority": {
                        "$ref": "#/components/schemas/TaskPriority"
                    },
                    "isComplete": {
                        "type": "boolean"
                    },
                    "dueAtUtc": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "isDeleted": {
                        "type": "boolean"
                    },
                    "orderIndex": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "projectId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "sectionId": {
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                    },
                    "assignedToUserId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "tagsIds": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "nullable": true
                    }
                },
                "additionalProperties": false
            },
...
            "UserProfileDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "email": {
                        "type": "string",
                        "nullable": true
                    },
                    "username": {
                        "type": "string",
                        "nullable": true
                    },
                    "displayName": {
                        "type": "string",
                        "nullable": true
                    },
                    "bio": {
                        "type": "string",
                        "nullable": true
                    },
                    "profilePictureUrl": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "additionalProperties": false
            }
        }
    }
}